Move RDP_DUMP over to Angrylion.
Change the dumping format to also support VI registers.
